+++ .cargo/config
@@ -5,7 +5,7 @@
 # It can only be run from 10x Genomics because it uses a bunch of data files that we don't
 # distribute.
 
-t = "test --bin traceback1 --test enclone_test1 --test enclone_test2 --test enclone_test3 --test enclone_test_peak_mem -- --nocapture"
+t = "test --bin traceback1 --test enclone_test1 --test enclone_test2 --test enclone_test3 --test enclone_test4 --test enclone_test_peak_mem -- --nocapture"
 
 # Note that "./test" exercises this plus a few more tests.
 
@@ -14,4 +14,4 @@ t = "test --bin traceback1 --test enclone_test1 --test enclone_test2 --test encl
 # "cargo tb" has to be run from the directory enclone_main, but should be runnable from any
 # machine, so long as at least the medium dataset has been downloaded.
 
-tb = "test --test enclone_test1 --test enclone_test2 --test enclone_test3 --features basic -- --nocapture"
+tb = "test --test enclone_test1 --test enclone_test2 --test enclone_test3 --test enclone_test4 --features basic -- --nocapture"
